+++ src/main/java/dynamicquad/agilehub/sprint/service/command/SprintService.java
@@ -11,10 +11,10 @@ import dynamicquad.agilehub.project.domain.Project;
 import dynamicquad.agilehub.project.service.MemberProjectService;
 import dynamicquad.agilehub.project.service.ProjectQueryService;
 import dynamicquad.agilehub.sprint.domain.Sprint;
-import dynamicquad.agilehub.sprint.domain.SprintRepository;
 import dynamicquad.agilehub.sprint.domain.SprintStatus;
 import dynamicquad.agilehub.sprint.dto.SprintRequestDto;
 import dynamicquad.agilehub.sprint.dto.SprintResponseDto;
+import dynamicquad.agilehub.sprint.repository.SprintRepository;
 import dynamicquad.agilehub.sprint.service.SprintValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
