+++ src/main/java/dynamicquad/agilehub/sprint/service/query/SprintQueryService.java
@@ -5,8 +5,8 @@ import dynamicquad.agilehub.issue.repository.IssueRepository;
 import dynamicquad.agilehub.member.dto.AssigneeDto;
 import dynamicquad.agilehub.project.service.ProjectQueryService;
 import dynamicquad.agilehub.sprint.domain.Sprint;
-import dynamicquad.agilehub.sprint.domain.SprintRepository;
 import dynamicquad.agilehub.sprint.dto.SprintResponseDto;
+import dynamicquad.agilehub.sprint.repository.SprintRepository;
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
